core: (fix?) initialization
WHY, WHY do we need this? After replacing the input-selector with an videomixer
the whole app upon starting behaved slow as a snail, BUT, the same pipeline when
run with gst-launch-1.0 was fine with more or less the same amount of cpu load.

Trial, error and the mighty stick revealed that doing a
[PLAYING->]READY->PLAYING state change fixed everything.