file archiving: don't leave unconnected queues.
Previously we had a tee with a spare queue for each element and we linked to that
whenever we wanted to dump to a file. However the queues would overflow
with different amount of data, leading to a/v sync issues (one track is
shorter)

Now we just request a new pad from the tee and use a couple of queues
internal to the MuxedFileWriter.