MainWindow: set handle for the main preview before adding inputs.
Otherwhise we risk it popping on its own window or crashing When adding
an input the pipeline goes to PLAYING, which in turn emits some time
later the prepare-window-handle signal, that grabs the xid from the
widget and then instructs the xvimagesink to draw there. But Gtk seems
to dislike being treated like that from different threads. At least that
is how I'm interpreting it now.