MainWindow: handling of preview widgets when adding sources.
Now TetraApp when adding a source returns us either the original source
or another one that wraps it (like when using a pair of inter elements),
so the PreviewWidget should point to that and not the source we
instantiated.

Also, instead of just removing the preview widget when a camera is gone
we leave it in place. We should give some kind of feedback about what
happened and when the camera is back again.